+++ projects/dsp-ui/src/lib/action/components/message/message.component.spec.ts
@@ -1,6 +1,5 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { Component, OnInit, ViewChild } from '@angular/core';
+import { async, ComponentFixture, fakeAsync, TestBed } from '@angular/core/testing';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { MatListModule } from '@angular/material/list';
@@ -11,10 +10,10 @@ import { StatusMsg } from '../../../../assets/i18n/statusMsg';
 import { DspMessageData, MessageComponent } from './message.component';
 
 /**
- * Test host component to simulate parent component with a progress bar.
+ * Test host component to simulate parent component.
  */
 @Component({
-    template: `<dsp-message #message [message]="shortMessage" [short]="short"></dsp-message>`
+    template: `<dsp-message #message [message]="shortMessage" [size]="size"></dsp-message>`
 })
 class ShortMessageTestHostComponent implements OnInit {
 
@@ -28,7 +27,7 @@ class ShortMessageTestHostComponent implements OnInit {
         footnote: 'Close it'
     };
 
-    short = true;
+    size = 'short';
 
     constructor() {
     }
@@ -37,10 +36,10 @@ class ShortMessageTestHostComponent implements OnInit {
 }
 
 /**
- * Test host component to simulate parent component with a progress bar.
+ * Test host component to simulate parent component.
  */
 @Component({
-    template: `<dsp-message #message [message]="errorMessage" [short]="short"></dsp-message>`
+    template: `<dsp-message #message [message]="errorMessage"></dsp-message>`
 })
 class LongMessageTestHostComponent implements OnInit {
 
@@ -53,7 +52,31 @@ class LongMessageTestHostComponent implements OnInit {
         error: 'error message'
     };
 
-    short = false;
+    constructor() {
+    }
+
+    ngOnInit() { }
+}
+
+/**
+ * Test host component to simulate parent component.
+ */
+@Component({
+    template: `<dsp-message #message [message]="shortMessage" [size]="size" [duration]="2000"></dsp-message>`
+})
+class ShortMessageWithDurationTestHostComponent implements OnInit {
+
+    @ViewChild('message', { static: false }) messageComponent: MessageComponent;
+
+    shortMessage: DspMessageData = {
+        status: 200,
+        statusMsg: 'Success',
+        statusText: 'You just updated the user profile.',
+        type: 'Note',
+        footnote: 'Close it'
+    };
+
+    size = 'short';
 
     constructor() {
     }
@@ -67,6 +90,9 @@ describe('MessageComponent', () => {
 
     let longMsgTestHostComponent: LongMessageTestHostComponent;
     let longMsgTestHostFixture: ComponentFixture<LongMessageTestHostComponent>;
+
+    let shortMsgDurationTestHostComponent: ShortMessageWithDurationTestHostComponent;
+    let shortMsgDurationTestHostFixture: ComponentFixture<ShortMessageWithDurationTestHostComponent>;
 
     let status: StatusMsg;
     let apiResonseError: ApiResponseError;
@@ -86,7 +112,8 @@ describe('MessageComponent', () => {
             declarations: [
                 MessageComponent,
                 ShortMessageTestHostComponent,
-                LongMessageTestHostComponent
+                LongMessageTestHostComponent,
+                ShortMessageWithDurationTestHostComponent
             ]
         }).compileComponents();
 
@@ -139,7 +166,8 @@ describe('MessageComponent', () => {
 
             expect(longMsgTestHostComponent.messageComponent.message.status).toEqual(403);
             expect(longMsgTestHostComponent.messageComponent.message.statusMsg).toEqual('Forbidden');
-            expect(longMsgTestHostComponent.messageComponent.message.statusText).toEqual('The request was a legal request, but the server is refusing to respond to it');
+            expect(longMsgTestHostComponent.messageComponent.message.statusText).toEqual(
+                'The request was a legal request, but the server is refusing to respond to it');
 
             const hostCompDe = longMsgTestHostFixture.debugElement;
 
@@ -151,8 +179,39 @@ describe('MessageComponent', () => {
 
             const messageTitleElement = messageEl.query(By.css('.message-title'));
 
-            expect(messageTitleElement.nativeElement.innerText).toEqual('The request was a legal request, but the server is refusing to respond to it');
+            expect(messageTitleElement.nativeElement.innerText).toEqual(
+                'The request was a legal request, but the server is refusing to respond to it');
 
         });
+    });
+
+    describe('display a short message with a duration of 2 seconds', () => {
+        beforeEach(() => {
+            shortMsgDurationTestHostFixture = TestBed.createComponent(ShortMessageWithDurationTestHostComponent);
+            shortMsgDurationTestHostComponent = shortMsgDurationTestHostFixture.componentInstance;
+            shortMsgDurationTestHostFixture.detectChanges();
+        });
+
+        it('should create', () => {
+            expect(shortMsgDurationTestHostComponent.messageComponent).toBeTruthy();
+        });
+
+        it('should display a short message', fakeAsync(() => {
+            expect(shortMsgDurationTestHostComponent.messageComponent).toBeTruthy();
+            expect(shortMsgDurationTestHostComponent.messageComponent.message.status).toEqual(200);
+            expect(shortMsgDurationTestHostComponent.messageComponent.message.statusMsg).toEqual('Success');
+
+            const hostCompDe = shortMsgDurationTestHostFixture.debugElement;
+
+            const messageEl = hostCompDe.query(By.directive(MessageComponent));
+
+            const spanShortMessageElement = messageEl.query(By.css('.dsp-short-message-text'));
+
+            expect(spanShortMessageElement.nativeElement.innerText).toEqual('You just updated the user profile.');
+
+            shortMsgDurationTestHostFixture.whenStable().then(() => {
+                expect(shortMsgDurationTestHostComponent.messageComponent.disable).toBeTruthy();
+            });
+        }));
     });
 });