+++ projects/dsp-ui/src/lib/viewer/operations/display-edit/display-edit.component.spec.ts
@@ -8,6 +8,8 @@ import { MatButtonHarness } from '@angular/material/button/testing';
 import { MAT_DIALOG_DATA, MatDialogModule, MatDialogRef } from '@angular/material/dialog';
 import { MatDialogHarness } from '@angular/material/dialog/testing';
 import { MatIconModule } from '@angular/material/icon';
+import { MatInputHarness } from '@angular/material/input/testing';
+import { MatTooltipModule } from '@angular/material/tooltip';
 import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import {
@@ -15,6 +17,7 @@ import {
     DeleteValue,
     DeleteValueResponse,
     MockResource,
+    MockUsers,
     ReadBooleanValue,
     ReadColorValue,
     ReadDecimalValue,
@@ -33,17 +36,22 @@ import {
     UpdateIntValue,
     UpdateResource,
     UpdateValue,
+    UsersEndpointAdmin,
     ValuesEndpointV2,
     WriteValueResponse
 } from '@dasch-swiss/dsp-js';
 import { of, throwError } from 'rxjs';
 import { AjaxError } from 'rxjs/ajax';
 import { DspApiConnectionToken } from '../../../core';
-import { EmitEvent, Events, ValueOperationEventService } from '../../services/value-operation-event.service';
+import {
+    DeletedEventValue,
+    EmitEvent,
+    Events,
+    UpdatedEventValues,
+    ValueOperationEventService
+} from '../../services/value-operation-event.service';
 import { ValueTypeService } from '../../services/value-type.service';
 import { DisplayEditComponent } from './display-edit.component';
-
-
 
 @Component({
   selector: `dsp-text-value-as-string`,
@@ -266,9 +274,12 @@ describe('DisplayEditComponent', () => {
   beforeEach(async(() => {
 
     const valuesSpyObj = {
-      v2: {
-        values: jasmine.createSpyObj('values', ['updateValue', 'getValue', 'deleteValue'])
-      }
+        admin: {
+            usersEndpoint: jasmine.createSpyObj('usersEndpoint', ['getUserByIri'])
+        },
+        v2: {
+            values: jasmine.createSpyObj('values', ['updateValue', 'getValue', 'deleteValue'])
+        }
     };
 
     const eventSpy = jasmine.createSpyObj('ValueOperationEventService', ['emit']);
@@ -277,7 +288,8 @@ describe('DisplayEditComponent', () => {
       imports: [
         BrowserAnimationsModule,
         MatIconModule,
-        MatDialogModule
+        MatDialogModule,
+        MatTooltipModule
       ],
       declarations: [
         DisplayEditComponent,
@@ -323,6 +335,17 @@ describe('DisplayEditComponent', () => {
   }));
 
   beforeEach(() => {
+
+    const adminSpy = TestBed.inject(DspApiConnectionToken);
+
+    // mock getUserByIri response
+    (adminSpy.admin.usersEndpoint as jasmine.SpyObj<UsersEndpointAdmin>).getUserByIri.and.callFake(
+        () => {
+            const user = MockUsers.mockUser();
+            return of(user);
+        }
+    );
+
     testHostFixture = TestBed.createComponent(TestHostDisplayValueComponent);
     testHostComponent = testHostFixture.componentInstance;
     testHostFixture.detectChanges();
@@ -490,7 +513,7 @@ describe('DisplayEditComponent', () => {
     it('should return the type of a integer value as not readonly', () => {
       expect(valueTypeService.getValueTypeOrClass(testHostComponent.displayEditValueComponent.displayValue)).toEqual(Constants.IntValue);
 
-      expect(valueTypeService.isReadOnly(Constants.IntValue)).toBe(false);
+      expect(valueTypeService.isReadOnly(Constants.IntValue, testHostComponent.displayEditValueComponent.displayValue)).toBe(false);
     });
 
     it('should return the class of a html text value as readonly', () => {
@@ -500,18 +523,7 @@ describe('DisplayEditComponent', () => {
 
       expect(valueTypeService.getValueTypeOrClass(htmlTextVal)).toEqual('ReadTextValueAsHtml');
 
-      expect(valueTypeService.isReadOnly('ReadTextValueAsHtml')).toBe(true);
-
-    });
-
-    it('should return the class of an XML text value as readonly', () => {
-
-      const xmlTextVal = new ReadTextValueAsXml();
-      xmlTextVal.type = Constants.TextValue;
-
-      expect(valueTypeService.getValueTypeOrClass(xmlTextVal)).toEqual('ReadTextValueAsXml');
-
-      expect(valueTypeService.isReadOnly('ReadTextValueAsXml')).toBe(true);
+      expect(valueTypeService.isReadOnly('ReadTextValueAsHtml', htmlTextVal)).toBe(true);
 
     });
 
@@ -522,7 +534,7 @@ describe('DisplayEditComponent', () => {
 
       expect(valueTypeService.getValueTypeOrClass(plainTextVal)).toEqual('ReadTextValueAsString');
 
-      expect(valueTypeService.isReadOnly('ReadTextValueAsString')).toBe(false);
+      expect(valueTypeService.isReadOnly('ReadTextValueAsString', plainTextVal)).toBe(false);
 
     });
 
@@ -578,79 +590,88 @@ describe('DisplayEditComponent', () => {
 
     it('should save a new version of a value', () => {
 
-      const valuesSpy = TestBed.inject(DspApiConnectionToken);
+        const valueEventSpy = TestBed.inject(ValueOperationEventService);
 
-      (valuesSpy.v2.values as jasmine.SpyObj<ValuesEndpointV2>).updateValue.and.callFake(
-        () => {
+        const valuesSpy = TestBed.inject(DspApiConnectionToken);
 
-          const response = new WriteValueResponse();
+        (valueEventSpy as jasmine.SpyObj<ValueOperationEventService>).emit.and.stub();
 
-          response.id = 'newID';
-          response.type = 'type';
-          response.uuid = 'uuid';
+        (valuesSpy.v2.values as jasmine.SpyObj<ValuesEndpointV2>).updateValue.and.callFake(
+            () => {
 
-          return of(response);
-        }
-      );
+                const response = new WriteValueResponse();
 
-      (valuesSpy.v2.values as jasmine.SpyObj<ValuesEndpointV2>).getValue.and.callFake(
-        () => {
+                response.id = 'newID';
+                response.type = 'type';
+                response.uuid = 'uuid';
 
-          const updatedVal = new ReadIntValue();
+                return of(response);
+            }
+        );
 
-          updatedVal.id = 'newID';
-          updatedVal.int = 1;
+        (valuesSpy.v2.values as jasmine.SpyObj<ValuesEndpointV2>).getValue.and.callFake(
+            () => {
 
-          const resource = new ReadResource();
+                const updatedVal = new ReadIntValue();
 
-          resource.properties = {
-            'http://0.0.0.0:3333/ontology/0001/anything/v2#hasInteger': [updatedVal]
-          };
+                updatedVal.id = 'newID';
+                updatedVal.int = 1;
 
-          return of(resource);
-        }
-      );
+                const resource = new ReadResource();
 
-      testHostComponent.displayEditValueComponent.canModify = true;
-      testHostComponent.displayEditValueComponent.editModeActive = true;
-      testHostComponent.displayEditValueComponent.mode = 'update';
+                resource.properties = {
+                    'http://0.0.0.0:3333/ontology/0001/anything/v2#hasInteger': [updatedVal]
+                };
 
-      testHostComponent.displayEditValueComponent.displayValueComponent.form.controls.test.clearValidators();
-      testHostComponent.displayEditValueComponent.displayValueComponent.form.controls.test.updateValueAndValidity();
+                return of(resource);
+            }
+        );
 
-      testHostFixture.detectChanges();
+        testHostComponent.displayEditValueComponent.canModify = true;
+        testHostComponent.displayEditValueComponent.editModeActive = true;
+        testHostComponent.displayEditValueComponent.mode = 'update';
 
-      const saveButtonDebugElement = displayEditComponentDe.query(By.css('button.save'));
-      const saveButtonNativeElement = saveButtonDebugElement.nativeElement;
+        testHostComponent.displayEditValueComponent.displayValueComponent.form.controls.test.clearValidators();
+        testHostComponent.displayEditValueComponent.displayValueComponent.form.controls.test.updateValueAndValidity();
 
-      expect(saveButtonNativeElement.disabled).toBeFalsy();
+        testHostFixture.detectChanges();
 
-      saveButtonNativeElement.click();
+        const saveButtonDebugElement = displayEditComponentDe.query(By.css('button.save'));
+        const saveButtonNativeElement = saveButtonDebugElement.nativeElement;
 
-      testHostFixture.detectChanges();
+        expect(saveButtonNativeElement.disabled).toBeFalsy();
 
-      const expectedUpdateResource = new UpdateResource();
+        saveButtonNativeElement.click();
 
-      expectedUpdateResource.id = 'http://rdfh.ch/0001/H6gBWUuJSuuO-CilHV8kQw';
-      expectedUpdateResource.type = 'http://0.0.0.0:3333/ontology/0001/anything/v2#Thing';
-      expectedUpdateResource.property = 'http://0.0.0.0:3333/ontology/0001/anything/v2#hasInteger';
+        testHostFixture.detectChanges();
 
-      const expectedUpdateVal = new UpdateIntValue();
-      expectedUpdateVal.id = 'http://rdfh.ch/0001/H6gBWUuJSuuO-CilHV8kQw/values/dJ1ES8QTQNepFKF5-EAqdg';
-      expectedUpdateVal.int = 1;
+        const expectedUpdateResource = new UpdateResource();
 
-      expectedUpdateResource.value = expectedUpdateVal;
+        expectedUpdateResource.id = 'http://rdfh.ch/0001/H6gBWUuJSuuO-CilHV8kQw';
+        expectedUpdateResource.type = 'http://0.0.0.0:3333/ontology/0001/anything/v2#Thing';
+        expectedUpdateResource.property = 'http://0.0.0.0:3333/ontology/0001/anything/v2#hasInteger';
 
-      expect(valuesSpy.v2.values.updateValue).toHaveBeenCalledWith(expectedUpdateResource);
-      expect(valuesSpy.v2.values.updateValue).toHaveBeenCalledTimes(1);
+        const expectedUpdateVal = new UpdateIntValue();
+        expectedUpdateVal.id = 'http://rdfh.ch/0001/H6gBWUuJSuuO-CilHV8kQw/values/dJ1ES8QTQNepFKF5-EAqdg';
+        expectedUpdateVal.int = 1;
 
-      expect(valuesSpy.v2.values.getValue).toHaveBeenCalledTimes(1);
-      expect(valuesSpy.v2.values.getValue).toHaveBeenCalledWith(testHostComponent.readResource.id,
-        'uuid');
+        expectedUpdateResource.value = expectedUpdateVal;
 
-      expect(testHostComponent.displayEditValueComponent.displayValue.id).toEqual('newID');
-      expect(testHostComponent.displayEditValueComponent.displayValueComponent.displayValue.id).toEqual('newID');
-      expect(testHostComponent.displayEditValueComponent.mode).toEqual('read');
+        expect(valuesSpy.v2.values.updateValue).toHaveBeenCalledWith(expectedUpdateResource);
+        expect(valuesSpy.v2.values.updateValue).toHaveBeenCalledTimes(1);
+
+        expect(valueEventSpy.emit).toHaveBeenCalledTimes(1);
+        expect(valueEventSpy.emit).toHaveBeenCalledWith(new EmitEvent(Events.ValueUpdated, new UpdatedEventValues(
+            testHostComponent.readValue, testHostComponent.displayEditValueComponent.displayValue)));
+
+        expect(valuesSpy.v2.values.getValue).toHaveBeenCalledTimes(1);
+        expect(valuesSpy.v2.values.getValue).toHaveBeenCalledWith(testHostComponent.readResource.id, 'uuid');
+
+        expect(testHostComponent.displayEditValueComponent.displayValue.id).toEqual('newID');
+        expect(testHostComponent.displayEditValueComponent.displayValueComponent.displayValue.id).toEqual('newID');
+        expect(testHostComponent.displayEditValueComponent.mode).toEqual('read');
+
+
 
     });
 
@@ -743,8 +764,6 @@ describe('DisplayEditComponent', () => {
     });
 
     it('should display a comment button if the value has a comment', () => {
-      //console.log(testHostComponent.displayEditValueComponent.displayValueComponent.displayValue);
-
       expect(testHostComponent.displayEditValueComponent.editModeActive).toBeFalsy();
       expect(testHostComponent.displayEditValueComponent.shouldShowCommentToggle).toBeTruthy()
 
@@ -888,15 +907,6 @@ describe('DisplayEditComponent', () => {
             }
         );
 
-        // const deleteButtonDebugElement = displayEditComponentDe.query(By.css('button.delete'));
-        // const deleteButtonNativeElement = deleteButtonDebugElement.nativeElement;
-
-        // expect(deleteButtonNativeElement.disabled).toBeFalsy();
-
-        // deleteButtonNativeElement.click();
-
-        // testHostFixture.detectChanges();
-
         const deleteButton = await rootLoader.getHarness(MatButtonHarness.with({selector: '.delete'}));
         await deleteButton.click();
 
@@ -917,6 +927,7 @@ describe('DisplayEditComponent', () => {
         const deleteVal = new DeleteValue();
         deleteVal.id = 'http://rdfh.ch/0001/H6gBWUuJSuuO-CilHV8kQw/values/dJ1ES8QTQNepFKF5-EAqdg';
         deleteVal.type = 'http://api.knora.org/ontology/knora-api/v2#IntValue';
+        deleteVal.deleteComment = undefined;
 
         expectedUpdateResource.value = deleteVal;
 
@@ -925,9 +936,48 @@ describe('DisplayEditComponent', () => {
             expect(valuesSpy.v2.values.deleteValue).toHaveBeenCalledTimes(1);
 
             expect(valueEventSpy.emit).toHaveBeenCalledTimes(1);
-            expect(valueEventSpy.emit).toHaveBeenCalledWith(new EmitEvent(Events.ValueDeleted, deleteVal));
+            expect(valueEventSpy.emit).toHaveBeenCalledWith(new EmitEvent(Events.ValueDeleted, new DeletedEventValue(deleteVal)));
         });
 
+    });
+
+    it('should send a deletion comment to Knora if one is provided', async () => {
+        const valueEventSpy = TestBed.inject(ValueOperationEventService);
+
+        const valuesSpy = TestBed.inject(DspApiConnectionToken);
+
+        (valueEventSpy as jasmine.SpyObj<ValueOperationEventService>).emit.and.stub();
+
+        (valuesSpy.v2.values as jasmine.SpyObj<ValuesEndpointV2>).deleteValue.and.callFake(
+            () => {
+
+                const response = new DeleteValueResponse();
+
+                response.result = 'success';
+
+                return of(response);
+            }
+        );
+
+        testHostComponent.displayEditValueComponent.deleteValue('my deletion comment');
+
+        const expectedUpdateResource = new UpdateResource();
+
+        expectedUpdateResource.id = 'http://rdfh.ch/0001/H6gBWUuJSuuO-CilHV8kQw';
+        expectedUpdateResource.type = 'http://0.0.0.0:3333/ontology/0001/anything/v2#Thing';
+        expectedUpdateResource.property = 'http://0.0.0.0:3333/ontology/0001/anything/v2#hasInteger';
+
+        const deleteVal = new DeleteValue();
+        deleteVal.id = 'http://rdfh.ch/0001/H6gBWUuJSuuO-CilHV8kQw/values/dJ1ES8QTQNepFKF5-EAqdg';
+        deleteVal.type = 'http://api.knora.org/ontology/knora-api/v2#IntValue';
+        deleteVal.deleteComment = 'my deletion comment';
+
+        expectedUpdateResource.value = deleteVal;
+
+        testHostFixture.whenStable().then(() => {
+            expect(valuesSpy.v2.values.deleteValue).toHaveBeenCalledWith(expectedUpdateResource);
+            expect(valuesSpy.v2.values.deleteValue).toHaveBeenCalledTimes(1);
+        });
     });
   });
 });