+++ projects/dsp-ui/src/lib/viewer/operations/display-edit/display-edit.component.ts
@@ -2,6 +2,7 @@ import { animate, state, style, transition, trigger } from '@angular/animations'
 import { Component, Inject, Input, OnInit, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import {
+    ApiResponseData,
     ApiResponseError,
     Constants,
     DeleteValue,
@@ -9,18 +10,27 @@ import {
     KnoraApiConnection,
     PermissionUtil,
     ReadResource,
+    ReadUser,
     ReadValue,
     UpdateResource,
     UpdateValue,
+    UserResponse,
     WriteValueResponse
 } from '@dasch-swiss/dsp-js';
 import { mergeMap } from 'rxjs/operators';
 import {
     ConfirmationDialogComponent,
-    ConfirmationDialogData
+    ConfirmationDialogData,
+    ConfirmationDialogValueDeletionPayload
 } from '../../../action/components/confirmation-dialog/confirmation-dialog.component';
 import { DspApiConnectionToken } from '../../../core/core.module';
-import { EmitEvent, Events, ValueOperationEventService } from '../../services/value-operation-event.service';
+import {
+    DeletedEventValue,
+    EmitEvent,
+    Events,
+    UpdatedEventValues,
+    ValueOperationEventService
+} from '../../services/value-operation-event.service';
 import { ValueTypeService } from '../../services/value-type.service';
 import { BaseValueComponent } from '../../values/base-value.component';
 
@@ -84,6 +94,14 @@ export class DisplayEditComponent implements OnInit {
     // string used as class name to add add to value-component element on hover
     backgroundColor = '';
 
+    dateDisplayOptions: 'era' | 'calendar' | 'all';
+
+    showDateLabels = false;
+
+    dateFormat: string;
+
+    user: ReadUser;
+
     constructor(
         @Inject(DspApiConnectionToken) private _dspApiConnection: KnoraApiConnection,
         private _valueOperationEventService: ValueOperationEventService,
@@ -94,6 +112,9 @@ export class DisplayEditComponent implements OnInit {
     ngOnInit() {
 
         this.mode = 'read';
+        this.dateDisplayOptions = 'all';
+        this.showDateLabels = true;
+        this.dateFormat = 'dd.MM.YYYY';
 
         // determine if user has modify permissions
         const allPermissions = PermissionUtil.allUserPermissions(this.displayValue.userHasPermission as 'RV' | 'V' | 'M' | 'D' | 'CR');
@@ -105,7 +126,21 @@ export class DisplayEditComponent implements OnInit {
 
         this.valueTypeOrClass = this._valueTypeService.getValueTypeOrClass(this.displayValue);
 
-        this.readOnlyValue = this._valueTypeService.isReadOnly(this.valueTypeOrClass);
+        this.readOnlyValue = this._valueTypeService.isReadOnly(this.valueTypeOrClass, this.displayValue);
+
+        this._dspApiConnection.admin.usersEndpoint.getUserByIri(this.displayValue.attachedToUser).subscribe(
+            (response: ApiResponseData<UserResponse>) => {
+                this.user = response.body.user;
+            },
+            (error: ApiResponseError) => {
+                console.error(error);
+            }
+        );
+    }
+
+    getTooltipText(): string {
+        return 'Creation date: ' + this.displayValue.valueCreationDate +
+            '\n Value creator: ' + this.user?.givenName + ' ' + this.user?.familyName;
     }
 
     /**
@@ -143,6 +178,10 @@ export class DisplayEditComponent implements OnInit {
                 })
             ).subscribe(
                 (res2: ReadResource) => {
+                    this._valueOperationEventService.emit(
+                        new EmitEvent(Events.ValueUpdated, new UpdatedEventValues(
+                            this.displayValue, res2.getValues(this.displayValue.property)[0])));
+
                     this.displayValue = res2.getValues(this.displayValue.property)[0];
                     this.mode = 'read';
 
@@ -177,18 +216,16 @@ export class DisplayEditComponent implements OnInit {
      */
     openDialog() {
         const dialogData = new ConfirmationDialogData();
-        dialogData.title = 'Are you sure want to delete this value from ' + this.displayValue.propertyLabel + '?';
-        dialogData.message = 'Confirming this action will delete the following value from ' +
-                                this.displayValue.propertyLabel + ':<br/><br/>' + this._generateValueInfo();
+        dialogData.value = this.displayValue;
         dialogData.buttonTextOk = 'Yes, delete the value';
         dialogData.buttonTextCancel = 'No, keep the value';
 
         const dialogRef =
             this._dialog.open<ConfirmationDialogComponent, ConfirmationDialogData>(ConfirmationDialogComponent, { data: dialogData});
 
-        dialogRef.afterClosed().subscribe((confirmed: boolean) => {
-            if (confirmed) {
-                this.deleteValue();
+        dialogRef.afterClosed().subscribe((payload: ConfirmationDialogValueDeletionPayload) => {
+            if (payload && payload.confirmed) {
+                this.deleteValue(payload.deletionComment);
             }
         });
     }
@@ -197,10 +234,11 @@ export class DisplayEditComponent implements OnInit {
      * Delete a value from a property.
      * Emits an event that can be listened to.
      */
-    deleteValue() {
+    deleteValue(comment?: string) {
         const deleteVal = new DeleteValue();
         deleteVal.id = this.displayValue.id;
         deleteVal.type = this.displayValue.type;
+        deleteVal.deleteComment = comment;
 
         const updateRes = new UpdateResource();
         updateRes.type = this.parentResource.type;
@@ -211,7 +249,7 @@ export class DisplayEditComponent implements OnInit {
         this._dspApiConnection.v2.values.deleteValue(updateRes as UpdateResource<DeleteValue>).pipe(
         mergeMap((res: DeleteValueResponse) => {
             // emit a ValueDeleted event to the listeners in resource-view component to trigger an update of the UI
-            this._valueOperationEventService.emit(new EmitEvent(Events.ValueDeleted, deleteVal));
+            this._valueOperationEventService.emit(new EmitEvent(Events.ValueDeleted, new DeletedEventValue(deleteVal)));
             return res.result;
         })).subscribe();
     }
@@ -254,9 +292,7 @@ export class DisplayEditComponent implements OnInit {
      * Show CRUD buttons and add 'highlighted' class to the element only if editModeActive is false
      */
     mouseEnter() {
-        if (this.canModify) {
-            this.showActionBubble = true;
-        }
+        this.showActionBubble = true;
         if (!this.editModeActive) {
             this.backgroundColor = 'highlighted';
         }
@@ -268,23 +304,6 @@ export class DisplayEditComponent implements OnInit {
     mouseLeave() {
         this.showActionBubble = false;
         this.backgroundColor = '';
-    }
-
-    /**
-     * Generate the message body for the confirmation dialog.
-     *
-     * @returns A string consisting of the values: value, comment, and creation date.
-     */
-    private _generateValueInfo(): string {
-        const value = this.displayValue.strval;
-        const comment = this.displayValue.valueHasComment ? this.displayValue.valueHasComment : 'No comment';
-        const creationDate = new Date(this.displayValue.valueCreationDate).toString();
-
-        const message = '<b>Value:</b> ' + value +
-                        '<br/><br/><b>Value Comment:</b> ' + comment +
-                        '<br/><br/><b>Value Creation Date:</b> ' + creationDate;
-
-        return message;
     }
 
 }