+++ src/app/action-playground/action-playground.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { ApiResponseError, StringLiteral } from '@dasch-swiss/dsp-js';
+import { ApiResponseError, ReadValue, StringLiteral } from '@dasch-swiss/dsp-js';
 import { ConfirmationDialogComponent, ConfirmationDialogData, DspMessageData, SortingService } from '@dasch-swiss/dsp-ui';
 
 @Component({
@@ -9,9 +9,6 @@ import { ConfirmationDialogComponent, ConfirmationDialogData, DspMessageData, So
   styleUrls: ['./action-playground.component.scss']
 })
 export class ActionPlaygroundComponent implements OnInit {
-
-    // loading: boolean;
-    // session: Session;
 
     sortProps: any = [
         {
@@ -132,19 +129,15 @@ export class ActionPlaygroundComponent implements OnInit {
     };
 
     confirmationDialogResponse: string;
+    showTimedMessage: boolean;
 
     constructor(
-        // private _sessionService: SessionService,
-        // @Inject(DspApiConnectionToken) private dspApiConnection: KnoraApiConnection,
         private _sortingService: SortingService,
         private _dialog: MatDialog
     ) { }
 
     ngOnInit(): void {
         this.refresh();
-
-        // already logged-in user?
-        // this.session = this._sessionService.getSession();
     }
 
     // only for testing the change of status
@@ -178,9 +171,14 @@ export class ActionPlaygroundComponent implements OnInit {
     // confirmation dialog
 
     openDialog() {
+        const testValue = new ReadValue();
+        testValue.strval = 'My data 101010101';
+        testValue.propertyLabel = 'My label';
+        testValue.valueCreationDate = '1993-10-10T19:11:00.00Z';
+        testValue.valueHasComment = 'My comment';
+
         const dialogData = new ConfirmationDialogData();
-        dialogData.title = 'Are you sure want to do this?';
-        dialogData.message = 'Confirming this action will delete the value. (Not really though, this is just a test message)';
+        dialogData.value = testValue;
         dialogData.buttonTextOk = 'Yes, delete the value';
         dialogData.buttonTextCancel = 'No, keep the value';
 
@@ -194,45 +192,11 @@ export class ActionPlaygroundComponent implements OnInit {
                 this.confirmationDialogResponse = 'Action was not confirmed';
             }
         });
-      }
-
-    // TODO: Will be replaced by login process from action module
-    /*
-    login() {
-        this.loading = true;
-        this.dspApiConnection.v2.auth.login('username', 'root', 'test').subscribe(
-            (response: ApiResponseData<LoginResponse>) => {
-                this._sessionService.setSession(response.body.token, 'root', 'username').subscribe(
-                    () => {
-                        this.loading = false;
-                        this.session = this._sessionService.getSession();
-                    });
-            },
-            (error: ApiResponseError) => {
-                // error handling
-                // this.loginErrorUser = (error.status === 404);
-                // this.loginErrorPw = (error.status === 401);
-                // this.loginErrorServer = (error.status === 0);
-
-                // this.errorMessage = error;
-
-                this.loading = false;
-                // TODO: update error handling similar to the old method (see commented code below)
-            }
-        );
     }
 
-    // TODO: Will be replaced by login process from action module
-    logout() {
-        this.loading = true;
-        this.dspApiConnection.v2.auth.logout().subscribe(
-            (response: ApiResponseData<LogoutResponse>) => {
-                this._sessionService.destroySession();
-                this.session = this._sessionService.getSession();
-                this.loading = false;
-            }
-        )
+    openMessage() {
+        this.showTimedMessage = true;
+        setTimeout(() => { this.showTimedMessage = false; }, 2100);
     }
-    */
 
 }