+++ src/app/app-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ActionPlaygroundComponent } from './action-playground/action-playground.component';
 import { AdvancedSearchPlaygroundComponent } from './advanced-search-playground/advanced-search-playground.component';
+import { CkeditorPlaygroundComponent } from './ckeditor/ckeditor-playground/ckeditor-playground.component';
 import { ModifyComponent } from './modify/modify.component';
 import { ReadComponent } from './read/read.component';
 import { SearchPlaygroundComponent } from './search-playground/search-playground.component';
@@ -31,7 +32,8 @@ const routes: Routes = [
     },
     { path: 'advanced-search', component: AdvancedSearchPlaygroundComponent}, // advanced search playground component
     { path: 'still-image', component: StillImagePlaygroundComponent}, // still image playground component
-    { path: 'viewer', component: ViewerPlaygroundComponent} // viewer playground component
+    { path: 'viewer', component: ViewerPlaygroundComponent}, // viewer playground component
+    { path: 'ckeditor', component: CkeditorPlaygroundComponent} // CKeditor playground component
 ];
 
 @NgModule({